[client] Break points (types + matching)
